Performance Improvements for Projectile Impact Checking
