+++ Orrery Combat Framework - Heart Module/Data/Scripts/HeartModule/Utility/CommandHandler.cs
@@ -30,6 +30,7 @@ namespace Heart_Module.Data.Scripts.HeartModule.Utility
             ["debug.fillammo"] = new Command("HeartMod.Debug", "Fills all magazines on your current grid.", (message) => I.FillGridWeapons()),
             ["debug.reloadammo"] = new Command("HeartMod.Debug", "Forces all weapons on your current grid to reload.", (message) => I.ReloadGridWeapons()),
             ["debug.reloaddefs"] = new Command("HeartMod.Debug", "Clears and refreshes all weapon definitions.", (message) => { HeartLoad.ResetDefinitions(); MyAPIGateway.Utilities.ShowMessage("[OCF]", "All definitions cleared. Good luck fixing the bug!"); }),
+            // TODO: Full on mod reload if possible
         };
 
         private void ShowHelp()
